:sparkles: Rename hooks, introduce environment info
