+++ cli/src/hooks/InitHook.ts
@@ -1,5 +1,5 @@
 import { Hook } from '@oclif/config';
-import { Collector } from './Collector';
+import { Collector } from '../Collector';
 
 const hook: Hook<'init'> = async function ({ id }) {
   const collector = new Collector(this.config);
